Extract Connection class
